+++ docs/guide/changelog.md
@@ -2,9 +2,9 @@
 
 <GlobalElement />
 
-## 1.10.2 <Tag color="volcano" size="small">soon</Tag>
+## 1.10.2 <Tag color="volcano" size="small">2024-12-05</Tag>
 
-- 优化并更新 [播放器 Video](https://themusecatcher.github.io/vue-amazing-ui/guide/components/video.html) 组件，新增 `iconSize` 属性；新增 `play` `pause` 方法；支持：设置播放暂停图标的尺寸；使用 `Methods` 控制播放暂停
+- 优化并更新 [播放器 Video](https://themusecatcher.github.io/vue-amazing-ui/guide/components/video.html) 组件，新增 `iconSize` 属性；重命名 `showPlay` 属性为 `playIcon`；新增 `play` `pause` 方法；支持：设置播放暂停图标的尺寸；使用 `Methods` 控制播放暂停
 
 ## 1.10.1 <Tag color="volcano" size="small">2024-12-04</Tag>